fofb_cc_clk_if.vhd, fofb_cc_gtp7_if: reformulate GTP7 reset
The main GTP reset (gtreset) is now directly connected to the system
reset and the user logic's reset is derived from the transceiver PLL
lock.

It tries to make the reset behavior similar to the other FPGA family
ports (Virtex II Pro, Virtex 5, Spartan 6 and Virtex 6), but using
a clocking scheme which is not based on MMCM/DCM/PLL. Instead of
waiting to the FPGA MMCM/DCM/PLL lock to deassert mgtreset, it uses
the transceiver PLL (Quad PLL) lock, which guarantees TXOUTCLK is
ready for use.